backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
pydantic-settings
pydantic[email]
mysqlclient
